Phase II/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database - SQLModel (SQLAlchemy + Pydantic)
sqlmodel==0.0.14
psycopg[binary]==3.1.18

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.2
